Added Admin logic

+++ src/main/java/softuniBlog/controller/ArticleController.java
@@ -2,6 +2,7 @@ package softuniBlog.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -50,6 +51,11 @@ public class ArticleController {
         if (!this.articleRepository.existsById(id)) {
         return "redirect:/";
         }
+        if(SecurityContextHolder.getContext().getAuthentication()instanceof AnonymousAuthenticationToken){
+        	UserDetails principal = (UserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        	User entityUser	 = this.userRepository.findByEmail(principal.getUsername());
+        	model.addAttribute("user", entityUser);
+        }
         Article article = this.articleRepository.getOne(id);
         model.addAttribute("article",article);
         model.addAttribute("view","article/details");
@@ -64,7 +70,9 @@ public class ArticleController {
                 return "redirect:/";
             }
             Article article = this.articleRepository.getOne(id);
-
+            if(!isUserAuthorOrAdmin(article)) {
+            	return "redirect:/article/" + id;
+            }
             model.addAttribute("view","article/edit");
             model.addAttribute("article",article);
             return "base-layout";
@@ -78,7 +86,9 @@ public class ArticleController {
                 return "redirect:/";
             }
             Article article = this.articleRepository.getOne(id);
-
+            if(!isUserAuthorOrAdmin(article)) {
+            	return "redirect:/article/" + id;
+            }
             article.setContent(articleBindingModel.getContent());
             article.setTitle(articleBindingModel.getTitle());
 
@@ -94,7 +104,9 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.getOne(id);
-
+        if(!isUserAuthorOrAdmin(article)) {
+        	return "redirect:/article/" + id;
+        }
         model.addAttribute("article",article);
         model.addAttribute("view", "article/delete");
         return "base-layout";
@@ -107,8 +119,19 @@ public String deleteProcess(@PathVariable Integer id){
             return "redirect:/";
         }
         Article article = this.articleRepository.getOne(id);
+        if(!isUserAuthorOrAdmin(article)) {
+        	return "redirect:/article/" + id;
+        }
+        
         this.articleRepository.delete(article);
+        
         return "redirect:/";
+}
+
+private boolean isUserAuthorOrAdmin(Article article) {
+	UserDetails user = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+	User userEntity = this.userRepository.findByEmail(user.getUsername());
+	return userEntity.isAdmin() || userEntity.isAuthor(article);
 }
 
 }
